restructures data/classes mods
