+++ py4DSTEM/classes/__init__.py
@@ -1,13 +1,7 @@
 _emd_hook = True
 
-from py4DSTEM.classes.calibration import Calibration
-from py4DSTEM.classes.data import Data
-from py4DSTEM.classes.datacube import DataCube
-from py4DSTEM.classes.diffractionslice import DiffractionSlice
-from py4DSTEM.classes.realslice import RealSlice
 from py4DSTEM.classes.probe import Probe
 from py4DSTEM.classes.virtualdiffraction import VirtualDiffraction
 from py4DSTEM.classes.virtualimage import VirtualImage
-from py4DSTEM.classes.qpoints import QPoints
