+++ py4DSTEM/classes/methods/__init__.py
@@ -1,3 +1,2 @@
-from py4DSTEM.classes.methods.datacube_methods import DataCubeMethods
 from py4DSTEM.classes.methods.probe_methods import ProbeMethods
