+++ py4DSTEM/data/datacube.py
@@ -1,17 +1,257 @@
-# Functions to become DataCube methods
+# Defines the DataCube class, which stores 4D-STEM datacubes
 
 import numpy as np
 from scipy.ndimage import distance_transform_edt, binary_fill_holes, gaussian_filter1d
 from scipy.interpolate import interp1d
+from typing import Optional,Union
 
-from emdfile import Array, Metadata, Node
+from emdfile import Array, Metadata, Node, Root
+from py4DSTEM.data import Data, Calibration
 
 
 
-class DataCubeMethods:
+
+class DataCube(Array,Data):
     """
-    A container for DataCube object instance methods.
+    Storage and processing methods for 4D-STEM datasets.
     """
+
+    def __init__(
+        self,
+        data: np.ndarray,
+        name: Optional[str] = 'datacube',
+        slicelabels: Optional[Union[bool,list]] = None,
+        setup_tree: Optional[bool] = True,
+        calibration: Optional[Union[Calibration,None]] = None,
+        ):
+        """
+        Accepts:
+            data (np.ndarray): the data
+            name (str): the name of the datacube
+            calibration (None or Calibration or 'pass'): default (None)
+                creates and attaches a new Calibration instance to root
+                metadata, or, passing a Calibration instance uses this instead.
+                'skip' is for internal use for the reader
+            slicelabels (None or list): names for slices if this is a
+                stack of datacubes
+
+        Returns:
+            A new DataCube instance.
+        """
+
+        # initialize DataCubeMethods
+        super(DataCubeMethods).__init__()
+
+        # initialize as an Array
+        Array.__init__(
+            self,
+            data = data,
+            name = name,
+            units = 'pixel intensity',
+            dim_names = [
+                'Rx',
+                'Ry',
+                'Qx',
+                'Qy'
+            ],
+            slicelabels = slicelabels
+        )
+
+        # initialize as Data
+        Data.__init__(
+            self,
+            setup_tree,
+            calibration
+        )
+
+        # cartesian coords
+        # TODO - tmp hack, needs to be refactored -
+        # this will break when preprocess methods are called
+        self.qyy,self.qxx = np.meshgrid(
+            np.arange(0,self.Q_Ny),
+            np.arange(0,self.Q_Nx)
+        )
+
+        # polar coords
+        self.polar = None
+
+
+
+    def _setup_calibration(self, cal):
+        """
+        Ensures that a calibration instance exists. Passing None
+        makes a new Calibration instance, puts it in root.calibration, and
+        makes a two way link. Passing a Calibration instance attaches that
+        instance. `'skip'` does nothing (internal use, on read from file).
+        """
+        if cal is None:
+            self.calibration = Calibration( datacube = self )
+        elif cal == 'skip':
+            pass
+        else:
+            assert(isinstance(cal, Calibration)), "`calibration` must be a Calibration instance, not type f{type(cal)}"
+            self.calibration = cal
+            cal._datacube = self
+
+
+    def copy(self):
+        """
+        Copys datacube
+        """
+        from py4DSTEM import DataCube
+        new_datacube = DataCube(
+            data = self.data.copy(),
+            name = self.name,
+            calibration = self.calibration.copy(),
+            slicelabels = self.slicelabels,
+        )
+
+        Qpixsize  = new_datacube.calibration.get_Q_pixel_size()
+        Qpixunits = new_datacube.calibration.get_Q_pixel_units()
+        Rpixsize  = new_datacube.calibration.get_R_pixel_size()
+        Rpixunits = new_datacube.calibration.get_R_pixel_units()
+
+        new_datacube.set_dim(
+            0,
+            [0,Rpixsize],
+            units = Rpixunits,
+            name = 'Rx'
+        )
+        new_datacube.set_dim(
+            1,
+            [0,Rpixsize],
+            units = Rpixunits,
+            name = 'Ry'
+        )
+
+        new_datacube.set_dim(
+            2,
+            [0,Qpixsize],
+            units = Qpixunits,
+            name = 'Qx'
+        )
+        new_datacube.set_dim(
+            3,
+            [0,Qpixsize],
+            units = Qpixunits,
+            name = 'Qy'
+        )
+
+        return new_datacube
+
+
+    # properties
+
+
+    ## pixel size / units
+
+    # Q 
+    @property
+    def Q_pixel_size(self):
+        return self.calibration.get_Q_pixel_size()
+    @property
+    def Q_pixel_units(self):
+        return self.calibration.get_Q_pixel_units()
+
+    # R
+    @property
+    def R_pixel_size(self):
+        return self.calibration.get_R_pixel_size()
+    @property
+    def R_pixel_units(self):
+        return self.calibration.get_R_pixel_units()
+
+    # aliases
+    qpixsize = Q_pixel_size
+    qpixunit = Q_pixel_units
+    rpixsize = R_pixel_size
+    rpixunit = R_pixel_units
+
+
+    ## shape
+
+    # FOV
+    @property
+    def R_Nx(self):
+        return self.data.shape[0]
+    @property
+    def R_Ny(self):
+        return self.data.shape[1]
+    @property
+    def Q_Nx(self):
+        return self.data.shape[2]
+    @property
+    def Q_Ny(self):
+        return self.data.shape[3]
+
+    @property
+    def Rshape(self):
+        return (self.data.shape[0],self.data.shape[1])
+    @property
+    def Qshape(self):
+        return (self.data.shape[2],self.data.shape[3])
+
+    @property
+    def R_N(self):
+        return self.R_Nx*self.R_Ny
+
+    # aliases
+    qnx = Q_Nx
+    qny = Q_Ny
+    rnx = R_Nx
+    rny = R_Ny
+    rshape = Rshape
+    qshape = Qshape
+    rn = R_N
+
+
+
+
+
+    # I/O
+
+    # to_h5 is inherited from Array
+
+    # read
+    @classmethod
+    def _get_constructor_args(cls,group):
+        """ Construct a datacube with no calibration / metadata
+        """
+        # We only need some of the Array constructors;
+        # dim vector/units are passed through when Calibration
+        # is loaded, and the runtim dim vectors are then set
+        # in _add_root_links
+        ar_args = Array._get_constructor_args(group)
+
+        args = {
+            'data': ar_args['data'],
+            'name': ar_args['name'],
+            'slicelabels': ar_args['slicelabels'],
+            'calibration': 'skip'
+        }
+
+        return args
+
+
+    def _add_root_links(self,group):
+        """ When reading from file, link to calibration metadata,
+        then use it to populate the datacube dim vectors
+        """
+        # Link to the datacube
+        self.calibration._datacube = self
+
+        # Populate dim vectors
+        self.calibration.set_Q_pixel_size( self.calibration.get_Q_pixel_size() )
+        self.calibration.set_R_pixel_size( self.calibration.get_R_pixel_size() )
+        self.calibration.set_Q_pixel_units( self.calibration.get_Q_pixel_units() )
+        self.calibration.set_R_pixel_units( self.calibration.get_R_pixel_units() )
+
+        return
+
+
+
+
+    # Class methods
 
     def add(
         self,
@@ -1664,8 +1904,6 @@ class DataCubeMethods:
             qr = np.hypot(self.qxx-vects.qx[idx], self.qyy-vects.qy[idx])
             mask = np.logical_and(mask, qr>radius)
         return mask
-
-
