+++ py4DSTEM/data/diffractionslice.py
@@ -2,7 +2,7 @@
 # diffraction-shaped data
 
 from emdfile import Array
-from py4DSTEM.classes import Data
+from py4DSTEM.data import Data
 
 from typing import Optional,Union
 import numpy as np
