+++ py4DSTEM/io/filereaders/empad.py
@@ -8,7 +8,7 @@
 import numpy as np
 from pathlib import Path
 from emdfile import tqdmnd
-from py4DSTEM.classes import DataCube
+from py4DSTEM.data import DataCube
 from py4DSTEM.preprocess.utils import bin2D
