+++ py4DSTEM/io/legacy/legacy12/read_v0_12.py
@@ -9,7 +9,7 @@ from emdfile import (
     PointList,
     PointListArray
 )
-from py4DSTEM.classes import (
+from py4DSTEM.data import (
     DataCube,
     DiffractionSlice,
     RealSlice,
