+++ py4DSTEM/process/latticevectors/fit.py
@@ -4,7 +4,7 @@ import numpy as np
 from numpy.linalg import lstsq
 
 from emdfile import tqdmnd, PointList, PointListArray
-from py4DSTEM.classes import RealSlice
+from py4DSTEM.data import RealSlice
 
 def fit_lattice_vectors(braggpeaks, x0=0, y0=0, minNumPeaks=5):
     """
