+++ py4DSTEM/visualize/vis_special.py
@@ -6,7 +6,7 @@ from mpl_toolkits.axes_grid1 import make_axes_locatable
 from scipy.spatial import Voronoi
 
 from emdfile import PointList
-from py4DSTEM.classes import DataCube,Calibration
+from py4DSTEM.data import DataCube,Calibration
 from py4DSTEM.process.utils import get_voronoi_vertices,convert_ellipse_params
 from py4DSTEM.process.calibration import double_sided_gaussian
 from py4DSTEM.process.latticevectors import get_selected_lattice_vectors
